Move require_me file to subdir
This keeps it out of the 'bad' sample list
