+++ samples/lib/require_me.rb
