+++ samples/simple_require.rb
@@ -1,8 +1,8 @@
 # frozen_string_literal: true
 #
-# This app tests that we are correctly setting load paths so files in the
-# executing directory can be required withour require_relative shenanigans.
-require 'require_me'
+# This app tests that we are correctly setting load paths so files relative to
+# the executing directory can be required withour require_relative shenanigans.
+require 'lib/require_me'
 
 Shoes.app do
   para RequireMe.message
